+++ src/main/kotlin/dev/jakubzika/worktracker/auth/SessionLogin.kt
@@ -2,4 +2,4 @@ package dev.jakubzika.worktracker.auth
 
 const val SESSION_LOGIN_NAME = "work-tracker-session-login"
 
-data class SessionLogin(val userId: Int, val userName: String)
+data class SessionLogin(val userId: Long, val userName: String)
